third_party/blink/web_tests/fast/css/border-shorthand-initialize-longhands.html

<!DOCTYPE html>
<html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
  <div id="target"></div>
  <script>
    test(() => {
      target.style.border = '1px';
      assert_equals(target.style.borderTopStyle, 'none');
      assert_equals(target.style.borderRightStyle, 'none');
      assert_equals(target.style.borderBottomStyle, 'none');
      assert_equals(target.style.borderLeftStyle, 'none');
      assert_equals(target.style.borderTopColor, 'currentcolor');
      assert_equals(target.style.borderRightColor, 'currentcolor');
      assert_equals(target.style.borderBottomColor, 'currentcolor');
      assert_equals(target.style.borderLeftColor, 'currentcolor');
      target.style.border = 'solid';
      assert_equals(target.style.borderTopWidth, 'medium');
      assert_equals(target.style.borderRightWidth, 'medium');
      assert_equals(target.style.borderBottomWidth, 'medium');
      assert_equals(target.style.borderLeftWidth, 'medium');
    }, 'Ensure border shorthand property initializes longhand properties such as border-top-style and border-right-color')
  </script>
</body>
</html>
